Speed-up ARM memcmp by leveraging NEON.
Because NEON_UNALIGNED_ACCESS has never been defined, it has gone unused.
This change changes it to check for __ARM_FEATURE_UNALIGNED, defined by GCC.
This has improved speed by roughly 34% on my testcases.

Also, memcmp checks if HAVE_32_BYTE_CACHE_LINE is defined and adjusts for it;
this is true for Cortex-A9, so define it in cortex-a9 makefile.

Change-Id: I3c76ce7fa2796872e0171d5502b0ebd6e2893339
Signed-off-by: ganeshi4u <ganeshvarmai4u@gmail.com>